move data loaders back to src
